+++ pkg/config/config_test.go
@@ -452,3 +452,28 @@ func TestLoadDynamicConfigurationExpandError(t *testing.T) {
 	assert.Error(t, err)
 	assert.True(t, strings.Contains(err.Error(), "expand var is not supported when using dynamic configuration, use gomplate env instead"))
 }
+
+func TestGRPCListenAddress(t *testing.T) {
+	cfgText := `
+metrics:
+  wal_directory: /tmp
+  scraping_service:
+    enabled: true
+    kvstore:
+      store: consul
+      consul: {}
+    lifecycler:
+      ring:
+        kvstore:
+          store: consul
+          consul: {}
+`
+	var c Config
+	err := LoadBytes([]byte(cfgText), false, &c)
+	require.NoError(t, err)
+	c.Server.Flags.GRPC.ListenAddress = "192.168.1.1:9090"
+	err = c.Validate(nil)
+	require.NoError(t, err)
+	require.Equal(t, 9090, c.Server.Flags.GRPC.ListenPort)
+	require.Equal(t, "192.168.1.1", c.Server.Flags.GRPC.ListenHost)
+}